+++ crates/fuel-core/src/service/sub_services/algorithm_updater.rs
@@ -200,6 +200,9 @@ pub fn get_synced_gas_price_updater(
             latest_block_height.into(),
             l2_block_source,
             metadata_storage,
+            config.min_gas_price,
+            config.gas_price_change_percent,
+            config.gas_price_threshold_percent,
         )
         .map_err(|e| anyhow::anyhow!("Could not initialize gas price updater: {e:?}"))
     }