+++ crates/fuel-core/src/service/adapters/txpool.rs
@@ -34,6 +34,7 @@ use fuel_core_types::{
         relayer::message::Message,
     },
     fuel_tx::{
+        BlobId,
         ConsensusParameters,
         Transaction,
         UtxoId,
@@ -42,6 +43,7 @@ use fuel_core_types::{
         ContractId,
         Nonce,
     },
+    fuel_vm::BlobData,
     services::{
         block_importer::SharedImportResult,
         p2p::{
@@ -133,6 +135,10 @@ impl fuel_core_txpool::ports::TxPoolDb for OnChainIterableKeyValueView {
 
     fn contract_exist(&self, contract_id: &ContractId) -> StorageResult<bool> {
         self.storage::<ContractsRawCode>().contains_key(contract_id)
+    }
+
+    fn blob_exist(&self, blob_id: &BlobId) -> StorageResult<bool> {
+        self.storage::<BlobData>().contains_key(blob_id)
     }
 
     fn message(&self, id: &Nonce) -> StorageResult<Option<Message>> {